Feedback::Handle Bug Assignment
* One bug can only be assigned to one Developer

* Once a bug is assigned to a developer, it cannot be reassigned to someone else

+++ app/controllers/project_assignments_controller.rb
@@ -5,19 +5,22 @@ class ProjectAssignmentsController < ApplicationController
 
   def create
     @project_assignment = ProjectAssignment.create(project_assignment_params)
+    authorize @project_assignment
 
     if @project_assignment.save
       AssignmentMailer.with(project: @project, user: @user).project_assignment_email.deliver_later
       redirect_to project_path(@project),
                   notice: "#{@user.name} was successfully added to #{@project.name}."
     else
-      redirect_to project_path(@project),
+      redirect_to users_path(user_type: @user_type, project_id: @project.id),
                   alert: "#{@user.name} has already been added to #{@project.name}."
     end
   end
 
   def destroy
     @project_assignment = ProjectAssignment.find_by(project_assignment_params)
+    authorize @project_assignment
+
     if @project_assignment.destroy
       AssignmentMailer.with(project: @project, user: @user).project_unassignment_email.deliver_later
       redirect_to project_path(@project),
@@ -37,5 +40,6 @@ class ProjectAssignmentsController < ApplicationController
   def set_params
     @project = Project.find(params[:project_id])
     @user = User.find(params[:user_id])
+    @user_type = params[:user_type]
   end
 end
